Changed dim_t and inc_t to be signed integers.
Details:
- Redefined dim_t and inc_t in terms of gint_t (instead of guint_t).
  This will facilitate interoperability with Fortran in the future.
  (Fortran does not support unsigned integers.)
- Redefined many instances of stride-related macros so that they return
  or use the absolute value of the strides, rather than the raw strides
  which may now be signed. Added new macros bli_is_row_stored_f() and
  bli_is_col_stored_f(), which assume positive (forward-oriented) strides,
  and changed the packm_blk_var[23] variants to use these macros instead
  of the existing bli_is_row_stored(), bli_is_col_stored().
- Added/adjusted typecasting to to various functions/macros, including
  bli_obj_alloc_buffer(), bli_obj_buffer_at_off(), and various pointer-
  related macros in bli_param_macro_defs.h.
- Redefined bli_convert_blas_incv() macro so that the BLAS compatibility
  layer properly handles situations where vector increments are negative.
  Thanks to Vladimir Sukharev for pointing out this issue.
- Changed type of increment parameters in bli_adjust_strides() from dim_t
  to inc_t. Likewise in bli_check_matrix_strides().
- Defined bli_check_matrix_object(), which checks for negative strides.
- Redefined bli_check_scalar_object() and bli_check_vector_object() so
  that they also check for negative stride.
- Added instances of bli_check_matrix_object() to various operations'
  _check routines.

+++ frame/base/bli_malloc.c
@@ -42,11 +42,11 @@ void* bli_malloc( siz_t size )
 	if ( size == 0 ) return NULL;
 
 #if BLIS_HEAP_ADDR_ALIGN_SIZE == 1
-	p = malloc( size );
+	p = malloc( ( size_t )size );
 #else
 	r_val = posix_memalign( &p,
-	                        ( size_t ) BLIS_HEAP_ADDR_ALIGN_SIZE,
-	                        ( size_t ) size );
+	                        ( size_t )BLIS_HEAP_ADDR_ALIGN_SIZE,
+	                        ( size_t )size );
 
 	if ( r_val != 0 ) bli_abort();
 #endif
